# Improve addon rendering/testing performance