… pkg/addon/types_test.go
… pkg/controller/seed-controller-manager/addon/addon_controller_test.go