**What this PR does / why we need it**:
We have a unit test, `TestRetagImageForAllVersions`, which is meant to ensure that

1. the code for loading and parsing addons works in itself, and
2. that all addons can be rendered for every Kubernetes version configured in KKP.

This works great, but is extremely slow. The slowness comes from the way we handled addon manifests, which were loaded from disk and then parsed as Go templates every time an addon needed to be rendered. This made the test so slow that we had to raise the `go test` timeout a few times.

This PR rewrites the addon parsing logic to keep the parsed Go templates permanently in memory. A new `addon.Addon` struct holds it and can be `.Render()`ed to produce runtime objects to be applied to user-clusters.

This means that addons are now loaded once upon startup of the seed-controller-manager. It speeds up its regular reconciliation loops, but means that if you are developing new addons using a local seed-ctrl-mgr, you will need to restart it every time an addon file is modified. In-cluster this won't be a problem, as the seed-ctrl-mgr Pod is recreated whenever the addon container image changes.
This then also means that bad addons will send the seed-ctrl-mgr into a crash loop.

These changes roughly halved the runtime for the unit test, from ~60 to 35 seconds.

Additionally I then also made it so we only test 1 version for each supported Kubernetes release. Previously every time we extended the list of versions in the default KubermaticConfiguration, it made the unit test take considerably longer. So a seemingly "cheap" action (adding 1 element to the version list) in reality costed us lots of time in the unit tests.

From experience addon manifests have never relied on the patch version, so I do not expect this change to realistically reduce the test coverage.

This change brought the runtime down from 35 to 10 seconds.

---

Additionally we recently hit #12739 and then learned that even though we tested addon rendering for all individual patch releaes for all CNIs, we only tested a small subset of cloud providers. Had we also rendered the addons for Digitalocean, we would have immediately noticed.

That's why this PR extends the unit tests to cover all providers and to ensure we never forget a new provider again. Adding these additional test rounds increased the runtime from 10 to roughly 20 seconds again.

**What type of PR is this?**
/kind cleanup

**Does this PR introduce a user-facing change? Then add your Release Note here**:
```release-note
Addon manifests are now loaded once upon startup of the seed-controller-manager instead of during every reconciliation. Invalid addons will now send the seed-controller-manager into a crash loop.
```

**Documentation**:
```documentation
NONE
```